requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
Django==4.1.7
django-bootstrap4==22.3
djangorestframework==3.14.0
pytz==2022.7.1
soupsieve==2.4
sqlparse==0.4.3
